refactor: implement GraphBuilder interface for graph parsing from yaml
- Introduced GraphBuilder interface to standardize graph creation and parsing.
- Added GraphBuilderYaml class for handling YAML code blocks in graphs.
- Updated main.ts to utilize GraphBuilderYaml for processing graph and graph3d code blocks.
- Refactored handleCodeBlock method to accept a GraphBuilder instance instead of a boolean flag.
- Removed unused methods and code from Utils class to streamline functionality.
